separate BeanCopier from KryoCloner.
